Address more review feedback.

+++ backend/src/main/java/com/redhat/cloud/notifications/routers/EndpointService.java
@@ -233,15 +233,17 @@ public class EndpointService {
         EndpointType endpointType = resources.getEndpointTypeById(principal.getAccount(), id);
         checkSystemEndpoint(endpointType);
 
-        Endpoint e = resources.getEndpoint(principal.getAccount(), id);
-        if (e != null) {
-            EndpointProperties properties = e.getProperties();
-            if (properties instanceof CamelProperties) {
-                CamelProperties cp = (CamelProperties) properties;
-                // Special case wrt OpenBridge
-                if (e.getSubType().equals("slack")) {
-                    String processorId = cp.getExtras().get(OB_PROCESSOR_ID);
-                    bridgeApiService.deleteProcessor(bridge.getId(), processorId, bridgeAuth.getToken());
+        if (obEnabled) {
+            Endpoint e = resources.getEndpoint(principal.getAccount(), id);
+            if (e != null) {
+                EndpointProperties properties = e.getProperties();
+                if (properties instanceof CamelProperties) {
+                    CamelProperties cp = (CamelProperties) properties;
+                    // Special case wrt OpenBridge
+                    if (e.getSubType().equals("slack")) {
+                        String processorId = cp.getExtras().get(OB_PROCESSOR_ID);
+                        bridgeApiService.deleteProcessor(bridge.getId(), processorId, bridgeAuth.getToken());
+                    }
                 }
             }
         }
